# openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "FixFlow API",
    "description": "Field service work order management API for dispatchers and technicians.",
    "version": "1"
  },
  "paths": {
...
    "/api/v1/parts": {
      "post": {
        "tags": [
          "Parts"
        ],
        "summary": "Create a part",
        "description": "Adds a part to the warehouse catalog with an initial stock quantity. The catalog number is trimmed, stored in upper case and must be unique, also among archived parts. Available to dispatchers and administrators.",
        "operationId": "CreatePart",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "headers": {
              "ETag": {
                "description": "Current version of the resource, to be sent in If-Match when changing it.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "get": {
        "tags": [
          "Parts"
        ],
        "summary": "List active parts",
        "description": "Returns one page of active parts with their current stock quantities, ordered by name and catalog number. Archived parts are not listed. The optional search matches a fragment of the name or catalog number regardless of letter case.",
        "operationId": "ListParts",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number, starting from 1.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Number of parts per page, from 1 to 100.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Case-insensitive fragment of the name or catalog number.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponseOfPartResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/parts/{partId}": {
      "get": {
        "tags": [
          "Parts"
        ],
        "summary": "Get a part",
        "description": "Returns a part with its current stock quantity by identifier, including archived parts so that historical service entries keep their part details.",
        "operationId": "GetPart",
        "parameters": [
          {
            "name": "partId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "ETag": {
                "description": "Current version of the resource, to be sent in If-Match when changing it.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "put": {
        "tags": [
          "Parts"
        ],
        "summary": "Update a part",
        "description": "Replaces the name, catalog number and unit price of an active part. The stock quantity is not changed and archived parts cannot be modified. A new price applies only to parts used afterwards. The If-Match header must carry the ETag of the part; a part changed in the meantime is rejected with 412. Available to dispatchers and administrators.",
        "operationId": "UpdatePart",
        "parameters": [
          {
            "name": "partId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "description": "ETag of the resource returned by the last read or change. A different current version is rejected with 412.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "headers": {
              "ETag": {
                "description": "Current version of the resource, to be sent in If-Match when changing it.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "412": {
            "description": "Precondition Failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "428": {
            "description": "Precondition Required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePartRequest": {
        "required": [
          "name",
          "catalogNumber",
          "stockQuantity",
          "unitPrice"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Part name."
          },
          "catalogNumber": {
            "type": "string",
            "description": "Catalog number, unique across all parts regardless of letter case and surrounding spaces."
          },
          "stockQuantity": {
            "type": "integer",
            "description": "Initial number of units in stock, from 0 to 100000.",
            "format": "int32"
          },
          "unitPrice": {
            "type": "number",
            "description": "Net unit price in PLN with at most two decimal places.",
            "format": "decimal"
          }
        },
        "description": "Data of a new part."
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "PagedResponseOfPartResponse": {
        "required": [
          "items",
          "page",
          "pageSize",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PartResponse"
            },
            "description": "Items on the requested page."
          },
          "page": {
            "type": "integer",
            "description": "Requested page number, starting from 1.",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "description": "Requested page size.",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "description": "Number of items on all pages.",
            "format": "int32"
          }
        },
        "description": "One page of a list."
      },
...
      "PartResponse": {
        "required": [
          "id",
          "name",
          "catalogNumber",
          "stockQuantity",
          "unitPrice",
          "createdAt",
          "archivedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identifier of the part.",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Part name."
          },
          "catalogNumber": {
            "type": "string",
            "description": "Unique catalog number, stored trimmed and in upper case."
          },
          "stockQuantity": {
            "type": "integer",
            "description": "Number of units currently in stock.",
            "format": "int32"
          },
          "unitPrice": {
            "type": "number",
            "description": "Current net unit price in PLN.",
            "format": "decimal"
          },
          "createdAt": {
            "type": "string",
            "description": "UTC time when the part was created.",
            "format": "date-time"
          },
          "archivedAt": {
            "type": [
              "null",
              "string"
            ],
            "description": "UTC time when the part was archived; null for active parts.",
            "format": "date-time"
          }
        },
        "description": "Spare part from the warehouse catalog."
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errorCode": {
            "type": [
              "null",
              "string"
            ],
            "description": "Stable machine-readable code of a business error, for example WorkOrder.NotFound. Absent for authentication and authorization failures."
          }
        },
        "description": "Error response in the RFC 9457 problem details format."
      },
...
      "UpdatePartRequest": {
        "required": [
          "name",
          "catalogNumber",
          "unitPrice"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Part name."
          },
          "catalogNumber": {
            "type": "string",
            "description": "Catalog number, unique across all parts regardless of letter case and surrounding spaces."
          },
          "unitPrice": {
            "type": "number",
            "description": "Net unit price in PLN with at most two decimal places.",
            "format": "decimal"
          }
        },
        "description": "New data of an existing part. All fields are replaced; the stock quantity changes only through deliveries and service entries."
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Access token returned by the login or refresh endpoint.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}